feat: added edit and delete methods for questions and answers

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRespository } from '../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -10,12 +11,17 @@ export class DeleteQuestionUseCase {
   constructor(private questionsRepository: QuestionsRespository) {}
 
   async execute({
+    authorId,
     questionId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId)
 
     if (!question) {
       throw new Error('Question not found')
+    }
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error('You are not allowed to delete this question')
     }
 
     await this.questionsRepository.delete(question)
